chore: mechanical restructure of classes
